Created class that contains all fromEntityToDto methods. Implemented everywhere.
