+++ src/main/java/com/cgm/infolab/controller/UserApiController.java
@@ -3,12 +3,12 @@ package com.cgm.infolab.controller;
 import com.cgm.infolab.db.model.UserEntity;
 import com.cgm.infolab.db.repository.UserRepository;
 import com.cgm.infolab.model.UserDto;
+import com.cgm.infolab.service.FromEntitiesToDtosService;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
 
-import java.security.Principal;
 import java.util.ArrayList;
 import java.util.List;
 
@@ -16,13 +16,14 @@ import java.util.List;
 public class UserApiController {
 
     private final UserRepository userRepository;
-
+    private final FromEntitiesToDtosService fromEntitiesToDtosService;
     private final Logger log = LoggerFactory.getLogger(UserApiController.class);
 
 
     @Autowired
-    public UserApiController(UserRepository userRepository) {
+    public UserApiController(UserRepository userRepository, FromEntitiesToDtosService fromEntitiesToDtosService) {
         this.userRepository = userRepository;
+        this.fromEntitiesToDtosService = fromEntitiesToDtosService;
     }
 
     @GetMapping("/api/users")
@@ -32,16 +33,10 @@ public class UserApiController {
         List<UserEntity> userEntities = userRepository.getByUsernameWithLike(user);
 
         if (userEntities.size() > 0) {
-
-            UserDtos = userEntities.stream().map(this::fromEntityToDto).toList();
+            UserDtos = userEntities.stream().map(fromEntitiesToDtosService::fromEntityToDto).toList();
         } else {
             log.info("Non sono stati trovati users");
         }
         return UserDtos;
-    }
-
-    public UserDto fromEntityToDto(UserEntity userEntity) {
-        UserDto userDto = UserDto.of(userEntity.getName().value(), userEntity.getId(), userEntity.getDescription());
-        return userDto;
     }
 }
